Refactor trajectory engine
