+++ analysis/trajectory_engine.py
